assignment_checker_project/requirements.txt
Django==5.2.6
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1

# PDF & OCR
pymupdf==1.23.8
pytesseract==0.3.10
Pillow==10.1.0

# Data Handling
pandas==2.1.4
openpyxl==3.1.2

setuptools<81
# LLM API
google-generativeai==0.8.3

python-dotenv==1.0.1
